backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database
PyMongo==4.6.0

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
Authlib==1.2.1

# Validation
pydantic==2.5.2
email-validator==2.2.0  # Required by pydantic for EmailStr

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0